refactor: Move diskwriter/router/connection to diskwriter/network/connection
